+++ Automate/Framework/Machines/Objects/MayonnaiseMachine.cs
@@ -82,7 +82,7 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
         /// <param name="location">The location containing the machine.</param>
         /// <param name="tile">The tile covered by the machine.</param>
         public MayonnaiseMachine(SObject machine, GameLocation location, Vector2 tile)
-            : base(machine, location, tile) { }
+            : base(machine, location, tile, machineTypeId: "MayonnaiseMachine") { }
 
         /// <summary>Provide input to the machine.</summary>
         /// <param name="input">The available items.</param>