+++ Automate/Framework/Models/ModConfig.cs
@@ -1,3 +1,6 @@
+using System.Collections.Generic;
+using Pathoschild.Stardew.Automate.Framework.Machines.Buildings;
+
 namespace Pathoschild.Stardew.Automate.Framework.Models
 {
     /// <summary>The raw mod configuration.</summary>
@@ -23,5 +26,11 @@ namespace Pathoschild.Stardew.Automate.Framework.Models
 
         /// <summary>Options affecting compatibility with other mods.</summary>
         public ModCompatibilityConfig ModCompatibility { get; set; } = new ModCompatibilityConfig();
+
+        /// <summary>The priority order in which to process machines. Machines have a default priority of 0, and higher values are processed first.</summary>
+        public IDictionary<string, int> MachinePriority = new Dictionary<string, int>
+        {
+            [nameof(ShippingBinMachine)] = -1
+        };
     }
 }